+++ src/eu/labrush/traveller/data/Point.java
@@ -1,5 +1,7 @@
 package eu.labrush.traveller.data;
 
+import java.math.BigInteger;
+
 public class Point {
     public int x ;
     public int y ;
@@ -9,9 +11,9 @@ public class Point {
         this.y = y ;
     }
 
-    public static int distance(Point p1, Point p2){
+    public static BigInteger distance(Point p1, Point p2){
         int dx = p1.x - p2.x ;
         int dy = p1.y - p2.y ;
-        return (int) Math.sqrt((double) (dx*dx + dy*dy)) ;
+        return BigInteger.valueOf((long) Math.sqrt((double) (dx*dx + dy*dy))) ;
     }
 }
